requirements.txt
# 华为云OBS SDK
huaweicloud-sdk-python-obs

# 环境变量管理
python-dotenv

# 定时任务
APScheduler

# 日志管理
loguru

# 进度条
tqdm
